Change how we deal with (implicit) -fno-rtti + -fsanitize=vptr
Summary:
We were trying to get cc1 to disable the vptr sanitizer, but cc1 only
looks at -fsanitize=, not at -fno-sanitize=

I'm changing that behavior (which is only relied on by us, AFAICT) to:
If we're on an “implicit -fno-rtti” platform and the vptr sanitizer is
enabled, warn that -frtti is being turned on for this compilation, and
turn it on.

I also refactored common tests and added -check-prefix=(NO-)?RTTI to some
RUN lines.

Reviewers: chandlerc, echristo

Subscribers: cfe-commits

Differential Revision: http://reviews.llvm.org/D7525